requirements.txt
django-extensions==0.8
ipython==0.12
wsgiref==0.1.2
